+++ app/discover/PubListing.tsx
@@ -4,6 +4,7 @@ import { usePubTheme } from "components/ThemeManager/PublicationThemeProvider";
 import { BaseThemeProvider } from "components/ThemeManager/ThemeProvider";
 import { PubLeafletPublication, PubLeafletThemeColor } from "lexicons/api";
 import { blobRefToSrc } from "src/utils/blobRefToSrc";
+import { timeAgo } from "src/utils/timeAgo";
 import { Json } from "supabase/database.types";
 
 export const PubListing = (props: {
@@ -69,26 +70,3 @@ export const PubListing = (props: {
     </BaseThemeProvider>
   );
 };
-
-export function timeAgo(timestamp: string): string {
-  const now = new Date();
-  const date = new Date(timestamp);
-  const diffMs = now.getTime() - date.getTime();
-  const diffSeconds = Math.floor(diffMs / 1000);
-  const diffMinutes = Math.floor(diffSeconds / 60);
-  const diffHours = Math.floor(diffMinutes / 60);
-  const diffDays = Math.floor(diffHours / 24);
-  const diffYears = Math.floor(diffDays / 365);
-
-  if (diffYears > 0) {
-    return `${diffYears} year${diffYears === 1 ? "" : "s"} ago`;
-  } else if (diffDays > 0) {
-    return `${diffDays} day${diffDays === 1 ? "" : "s"} ago`;
-  } else if (diffHours > 0) {
-    return `${diffHours} hour${diffHours === 1 ? "" : "s"} ago`;
-  } else if (diffMinutes > 0) {
-    return `${diffMinutes} minute${diffMinutes === 1 ? "" : "s"} ago`;
-  } else {
-    return "just now";
-  }
-}
